% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty.R
\name{Befun}
\alias{Befun}
\title{Beta function integrand for penalty calculation}
\usage{
Befun(w, al = al, bl = bl, p = p, q = q)
}
\arguments{
\item{w}{Integration variable.}

\item{al}{Lower bound.}

\item{bl}{Upper bound.}

\item{p}{Shape parameter p.}

\item{q}{Shape parameter q.}
}
\value{
Integrand value.
}
\description{
Beta function integrand for penalty calculation
}
\keyword{internal}
