% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev.R
\name{boot.cov}
\alias{boot.cov}
\title{Bootstrap covariance of sample L-moments}
\usage{
boot.cov(xdat, BB)
}
\arguments{
\item{xdat}{Numeric data vector.}

\item{BB}{Number of bootstrap replicates.}
}
\value{
A 3x3 covariance matrix.
}
\description{
Computes the covariance matrix of the first three sample
L-moments by nonparametric bootstrap. Used as a fallback when the
direct covariance estimate is singular.
}
\keyword{internal}
