% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{calc_time_m_11}
\alias{calc_time_m_11}
\title{Time-varying moment estimation}
\usage{
calc_time_m_11(qmax = NULL, orig.para = NULL, rob = FALSE, mdfy = FALSE)
}
\description{
Time-varying moment estimation
}
\keyword{internal}
