% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magev_helpers.R
\name{cd.hos}
\alias{cd.hos}
\title{Coles-Dixon Penalty Function}
\usage{
cd.hos(sxi)
}
\arguments{
\item{sxi}{Shape parameter value (Hosking style, negative for heavy tails).}
}
\value{
Prior probability value.
}
\description{
Computes the Coles-Dixon prior probability for the shape parameter.
}
\keyword{internal}
