% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magev_helpers.R
\name{cov.dir}
\alias{cov.dir}
\title{Dirichlet covariance matrix for weights}
\usage{
cov.dir(wtgd)
}
\arguments{
\item{wtgd}{Numeric vector of model weights.}
}
\value{
A (numk x numk) covariance matrix.
}
\description{
Internal function that computes the covariance matrix of
weights assuming a Dirichlet-type distribution.
}
\keyword{internal}
