% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magev_helpers.R
\name{emp.prior}
\alias{emp.prior}
\title{Empirical prior for BMA based on MLE and LME}
\usage{
emp.prior(mle3, mle, lme)
}
\arguments{
\item{mle3}{Matrix of candidate submodel parameters (numk x 3).}

\item{mle}{MLE parameter estimates (mu, sigma, xi).}

\item{lme}{LME parameter estimates (mu, sigma, xi).}
}
\value{
Numeric vector of prior probability values for each candidate.
}
\description{
Internal function that computes an empirical prior distribution
for the shape parameter based on the range between MLE and LME estimates.
}
\keyword{internal}
