% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{find_max_beta.pk}
\alias{find_max_beta.pk}
\title{Find the shape parameter maximizing the beta penalty}
\usage{
find_max_beta.pk(lme.center, p = p, c0 = c0, c1 = c1, c2 = c2)
}
\arguments{
\item{lme.center}{Center vector (mu0, sigma0, xi) of the penalty.}

\item{p, c0, c1, c2}{Beta penalty hyperparameters.}
}
\value{
The xi value maximizing the penalty.
}
\description{
Finds the value of xi (over a grid within the adaptive
beta penalty support) which maximizes the beta penalty function.
Used as one of the initial points for optimization.
}
\keyword{internal}
