% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{fun.lme.gev11}
\alias{fun.lme.gev11}
\title{L-moment distance function for GEV11 model}
\usage{
fun.lme.gev11(a, xdat = xdat, pretheta = pretheta)
}
\arguments{
\item{a}{Numeric vector (mu0, sigma0, xi) to optimize.}

\item{xdat}{Numeric vector of data.}

\item{pretheta}{Full parameter vector (mu0, mu1, sigma0, sigma1, xi)
from pre-estimation; mu1 and sigma1 are held fixed.}
}
\value{
Numeric vector of length 3 of L-moment equation residuals.
}
\description{
Computes the vector of differences between the population
L-moments of the standard Gumbel distribution and the sample L-moments
of the data transformed by \code{\link{trans.gum01}} under the current
GEV11 parameters. Solving this system equal to zero gives the L-moment
estimates (Shin et al., 2025b).
}
\keyword{internal}
