% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{gev.Ldist.m0s0_11}
\alias{gev.Ldist.m0s0_11}
\title{L-moment distance function for GEV11 model}
\usage{
gev.Ldist.m0s0_11(a, xdat = xdat, pretheta = pretheta)
}
\description{
L-moment distance function for GEV11 model
}
\keyword{internal}
