% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magev_helpers.R
\name{ginit.xifix}
\alias{ginit.xifix}
\title{Initialize parameters for GEV with fixed xi}
\usage{
ginit.xifix(data, ntry)
}
\arguments{
\item{data}{Numeric vector of data.}

\item{ntry}{Number of initial parameter sets to generate.}
}
\value{
A matrix with \code{ntry} rows and 2 columns (mu, sigma).
}
\description{
Internal function that generates initial (mu, sigma) values
for optimization with fixed shape parameter, using Gumbel L-moment estimates
and random perturbations.
}
\keyword{internal}
