% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{gof.ene_all}
\alias{gof.ene_all}
\title{Goodness-of-fit function}
\usage{
gof.ene_all(xdat, vecT = c(5, 10, 20, 40, 80), para, model = NULL)
}
\description{
Goodness-of-fit function
}
\keyword{internal}
