% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haenam}
\alias{haenam}
\title{Haenam Maximum Rainfall Data}
\format{
A data frame containing annual maximum daily rainfall values.
}
\source{
Korea Meteorological Administration (KMA; \url{https://www.kma.go.kr})
}
\usage{
haenam
}
\description{
Annual maximum daily rainfall data from Haenam, South Korea.
This dataset is used for demonstrating model averaging methods
for high quantile estimation in extreme value analysis.
}
\examples{
data(haenam)
head(haenam)

# Estimate high quantiles using model averaging
\donttest{
result <- ma.gev(haenam[,1], quant = c(0.98, 0.99, 0.995))
print(result$qua.ma)
}

}
\references{
Shin, Y., Shin, Y., & Park, J. S. (2026). Model averaging with mixed criteria
for estimating high quantiles of extreme values: Application to heavy rainfall.
\emph{Stochastic Environmental Research and Risk Assessment}, 40(2), 47.
\doi{10.1007/s00477-025-03167-x}
}
\keyword{datasets}
