% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magev_helpers.R
\name{is.odd.me}
\alias{is.odd.me}
\title{Check if a number is odd}
\usage{
is.odd.me(x)
}
\arguments{
\item{x}{Integer value.}
}
\value{
Logical. TRUE if x is odd, FALSE otherwise.
}
\description{
Check if a number is odd
}
\keyword{internal}
