% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{lmoms.md.park}
\alias{lmoms.md.park}
\title{Modified L-moments calculation}
\usage{
lmoms.md.park(x, nmom = 5, mtrim = FALSE, no.stop = FALSE, vecit = FALSE)
}
\description{
Modified L-moments calculation
}
\keyword{internal}
