% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{multi.m0s0_11}
\alias{multi.m0s0_11}
\title{Multi-start optimization for GEV11 model}
\usage{
multi.m0s0_11(
  xdat,
  ntry = 10,
  ftol = 1e-06,
  pretheta = pretheta,
  model = model,
  pen = "beta",
  mu = mu,
  p = p,
  std = std,
  c1 = c1,
  c2 = c2,
  choose = "gof"
)
}
\description{
Multi-start optimization for GEV11 model
}
\keyword{internal}
