% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{obj.lme.gev11}
\alias{obj.lme.gev11}
\title{L-moment distance function for GEV11 model}
\usage{
obj.lme.gev11(a, xdat = xdat, pretheta = pretheta)
}
\arguments{
\item{a}{Numeric vector (mu0, sigma0, xi) to optimize.}

\item{xdat}{Numeric vector of data.}

\item{pretheta}{Full parameter vector (mu0, mu1, sigma0, sigma1, xi) from pre-estimation.}
}
\value{
Numeric vector of length 3, representing L-moment equation residuals.
}
\description{
Internal function that computes the L-moment distance vector
for the GEV11 model. The residuals are formed from the difference
between theoretical Gumbel L-moments and sample L-moments of
standardized data.
}
\keyword{internal}
