% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev.R
\name{pargev.kfix}
\alias{pargev.kfix}
\title{GEV parameter estimation with fixed shape parameter}
\usage{
pargev.kfix(lmom, kfix = 0.1, checklmom = TRUE, ...)
}
\arguments{
\item{lmom}{L-moments object.}

\item{kfix}{Fixed shape parameter value.}

\item{checklmom}{Whether to check L-moment validity.}

\item{...}{Additional arguments.}
}
\value{
A list with GEV parameters.
}
\description{
GEV parameter estimation with fixed shape parameter
}
\author{
Jeong-Soo Park
}
