% Generated by roxygen2: do not edit by hand
\name{PhliuAgromet}
\alias{PhliuAgromet}
\docType{data}
\title{Phliu Agrometeorological Station Data}
\description{
Climate or meteorological data from the Phliu Agrometeorological Station.
}
\usage{data(PhliuAgromet)}
\format{
A data frame containing meteorological measurements.
}
\source{
Phliu Agrometeorological Station.
}
\keyword{datasets}
