% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev.R
\name{pk.beta.stnary}
\alias{pk.beta.stnary}
\title{Beta preference function for stationary GEV}
\usage{
pk.beta.stnary(para = NULL, lme.center = NULL, p = NULL, c0 = 0.3, c1 = 10, c2 = 5)
}
\arguments{
\item{para}{A vector of GEV parameters.}

\item{lme.center}{L-moment estimates as center.}

\item{p}{Shape parameter (default 6).}

\item{c0}{Limit parameter (default 0.3).}

\item{c1}{Scaling parameter (default 10).}

\item{c2}{Upper limit parameter (default 5).}
}
\value{
A list containing pk.one (preference value), p, and q.
}
\description{
Calculates a beta preference function for the stationary GEV shape parameter.
}
\author{
Jeong-Soo Park
}
