% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{pk.norm.ns}
\alias{pk.norm.ns}
\title{Normal preference function for non-stationary GEV}
\usage{
pk.norm.ns(para = NULL, mu = NULL, std = NULL)
}
\arguments{
\item{para}{A vector of GEV parameters.}

\item{mu}{Mean for normal distribution.}

\item{std}{Standard deviation for normal distribution.}
}
\value{
Preference function value.
}
\description{
Normal preference function for non-stationary GEV
}
\author{
Jeong-Soo Park
}
\keyword{internal}
