% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magev_helpers.R
\name{PrescottW}
\alias{PrescottW}
\title{Prescott-Walden expected information matrix for GEV with fixed xi}
\usage{
PrescottW(par2 = NULL, xifix = NULL, nsam = NULL)
}
\arguments{
\item{par2}{Numeric vector of (mu, sigma) estimates.}

\item{xifix}{Fixed shape parameter value.}

\item{nsam}{Sample size.}
}
\value{
A 2x2 expected information matrix.
}
\description{
Internal function that computes the Prescott-Walden expected
information (Hessian) matrix for GEV parameters (mu, sigma) with fixed xi.
}
\keyword{internal}
