% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magev_helpers.R
\name{prior.beta.stnary.ma}
\alias{prior.beta.stnary.ma}
\title{Beta prior for model averaging}
\usage{
prior.beta.stnary.ma(x = NULL, xi_lme = NULL, p = 2, q = 5, al, bl)
}
\arguments{
\item{x}{Shape parameter value to evaluate.}

\item{xi_lme}{L-moment estimate of xi (unused in this version).}

\item{p}{Beta shape parameter p. Default is 2.}

\item{q}{Beta shape parameter q. Default is 5.}

\item{al}{Lower bound of support.}

\item{bl}{Upper bound of support.}
}
\value{
Prior probability value (scalar).
}
\description{
Internal function that computes a Beta distribution prior
probability for a candidate shape parameter in model averaging.
}
\keyword{internal}
