% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran.gev.R
\name{qns.gev11}
\alias{qns.gev11}
\title{Quantile function for GEV11 model}
\usage{
qns.gev11(Tp = NULL, para = NULL, year = NULL)
}
\arguments{
\item{Tp}{Return period.}

\item{para}{Parameter vector (mu0, mu1, sigma0, sigma1, xi).}

\item{year}{Time vector.}
}
\value{
Numeric vector of quantile values (one per time point).
}
\description{
Internal wrapper that computes time-varying quantiles for the
GEV11 model given a return period.
}
\keyword{internal}
