% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{qns.gev_all}
\alias{qns.gev_all}
\title{Quantile function for non-stationary GEV}
\usage{
qns.gev_all(Tp = NULL, para, year, model = NULL)
}
\description{
Quantile function for non-stationary GEV
}
\keyword{internal}
