% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran.gev.R
\name{ran.gev_all}
\alias{ran.gev_all}
\title{Random number generation for non-stationary GEV models}
\usage{
ran.gev_all(nsample, para, model = NULL)
}
\arguments{
\item{nsample}{Sample size (number of time points).}

\item{para}{Parameter vector; its length and meaning depend on
\code{model} (see \code{\link{quagev.NS}}). For "gev11":
(mu0, mu1, sigma0, sigma1, xi).}

\item{model}{Model type: "gev", "gev00", "gev01", "gev10", "gev11",
or "gev20".}
}
\value{
A numeric vector of length \code{nsample}.
}
\description{
Generates a random sample from a stationary or non-stationary GEV model.
For non-stationary models the location and/or scale vary with time
t = 1, ..., nsample:
mu(t) = mu0 + mu1*t + mu2*t^2 and sigma(t) = exp(sigma0 + sigma1*t).
}
\examples{
set.seed(1)
# Non-stationary GEV11 sample of size 50
x <- ran.gev_all(50, para = c(100, 0.5, 3, 0.005, -0.2), model = "gev11")
plot(x, type = "l")

}
\references{
Shin, Y., Shin, Y. & Park, J.-S. (2025). Building nonstationary extreme value
model using L-moments. Journal of the Korean Statistical Society, 54, 947-970.
\doi{10.1007/s42952-025-00325-3}
}
\seealso{
\code{\link{quagev.NS}} for the corresponding quantile function,
  \code{\link{glme.gev11}}, \code{\link{lme.gev11}}.
}
\author{
Yonggwan Shin, Seokkap Ko, Jihong Park, Yire Shin, Jeong-Soo Park
}
