% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magev_helpers.R
\name{rcd}
\alias{rcd}
\title{Revised Coles-Dixon prior function}
\usage{
rcd(xi)
}
\arguments{
\item{xi}{Numeric vector of shape parameter values.}
}
\value{
Numeric vector of prior probability values.
}
\description{
Internal function that computes a revised Coles-Dixon prior
probability for the shape parameter, used as a penalty in BMA.
}
\keyword{internal}
