% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streamflow}
\alias{streamflow}
\title{Streamflow Data}
\format{
A data frame with 50 rows and 2 columns:
\describe{
  \item{Year}{Year of observation (character)}
  \item{r1}{Annual maximum streamflow value (numeric)}
}
}
\source{
Hydrological data for extreme value analysis.
}
\usage{
streamflow
}
\description{
Annual maximum streamflow measurements for extreme value analysis.
}
\examples{
data(streamflow)
head(streamflow)

}
\keyword{datasets}
