% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{strup_11_glme_const}
\alias{strup_11_glme_const}
\title{Strup WLS with ordinary regression}
\usage{
strup_11_glme_const(xdat, orig.para = orig.para, const = NULL)
}
\description{
Strup WLS with ordinary regression
}
\keyword{internal}
