% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{strup_11_glme_const_lmrob}
\alias{strup_11_glme_const_lmrob}
\title{Strup WLS with robust regression}
\usage{
strup_11_glme_const_lmrob(xdat, orig.para = orig.para, const = NULL)
}
\description{
Strup WLS with robust regression
}
\keyword{internal}
