% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GN16.gev11.R
\name{time.m.gev11}
\alias{time.m.gev11}
\title{Time-varying moment step of the GN16 method}
\usage{
\method{time}{m.gev11}(qmax = NULL, orig.para = NULL, rob = FALSE)
}
\arguments{
\item{qmax}{Pseudo-maxima series from \code{\link{make.qmax.gev11}}.}

\item{orig.para}{Parameter vector (mu0, mu1, sigma0, sigma1, xi).}

\item{rob}{If TRUE, uses robust regression for the location trend.}
}
\value{
A list with para.gado.org and para.gado.mdfy.
}
\description{
Internal step of \code{\link{GN16.gev11}}: fits a
stationary GEV to the pseudo-maxima and converts the result to the
GEV11 parameterization (both original and modified specification).
}
\keyword{internal}
