% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{trans.gum01}
\alias{trans.gum01}
\title{Transform non-stationary GEV data to standard Gumbel}
\usage{
trans.gum01(xdat, para = NULL, na.ok = FALSE)
}
\arguments{
\item{xdat}{Numeric vector of data.}

\item{para}{GEV11 parameter vector (mu0, mu1, sigma0, sigma1, xi).}

\item{na.ok}{If TRUE, invalid (out-of-support) values are returned as NA;
if FALSE, they are clipped.}
}
\value{
Numeric vector on the standard Gumbel scale.
}
\description{
Transforms observations from a non-stationary GEV11 model
with parameters \code{para} = (mu0, mu1, sigma0, sigma1, xi) to the
standard Gumbel scale.
}
\keyword{internal}
