% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strup.gev11.R
\name{wls.gev11}
\alias{wls.gev11}
\title{Log-scale regression step of the Strupczewski WLS method}
\usage{
wls.gev11(xdat, res = NULL, rob = NULL)
}
\arguments{
\item{xdat}{Numeric data vector.}

\item{res}{Residuals from the location trend fit.}

\item{rob}{If TRUE, uses robust regression (setting "KS2014").}
}
\value{
A list with element \code{sig} (the two coefficients).
}
\description{
Internal step of \code{\link{strup.gev11}}: regresses the
log absolute residuals on time to estimate the log-scale coefficients
(sigma0, sigma1).
}
\keyword{internal}
