% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GN16.gev11.R
\name{GN16.gev11}
\alias{GN16.gev11}
\title{GN16 estimation for the non-stationary GEV11 model}
\usage{
GN16.gev11(xdat, rob = FALSE)
}
\arguments{
\item{xdat}{A numeric vector of data to be fitted.}

\item{rob}{If TRUE, uses robust regression in the internal steps
(default FALSE).}
}
\value{
A list containing:
\itemize{
  \item para.gado.org - Original GN16 estimates
    (mu0, mu1, sigma0, sigma1, xi).
  \item para.gado.mdfy - Modified GN16 estimates.
}
}
\description{
Estimates the parameters of the non-stationary GEV11 model by the
quantile-based GN16 method (Gilleland and Katz, 2016; as adapted in
Shin et al., 2025b), with both the original and the modified
specification returned.
}
\examples{
data(PhliuAgromet)
result <- GN16.gev11(PhliuAgromet$prec)
print(result$para.gado.org)

}
\references{
Gilleland, E. & Katz, R. W. (2016). extRemes 2.0: An extreme value
analysis package in R. Journal of Statistical Software, 72(8), 1-39.

Shin, Y., Shin, Y. & Park, J.-S. (2025). Building nonstationary extreme value
model using L-moments. Journal of the Korean Statistical Society, 54, 947-970.
\doi{10.1007/s42952-025-00325-3}
}
\seealso{
\code{\link{glme.gev11}}, \code{\link{lme.gev11}},
  \code{\link{strup.gev11}}.
}
\author{
Yonggwan Shin, Seokkap Ko, Jihong Park, Yire Shin, Jeong-Soo Park
}
