% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{gof.ene_all}
\alias{gof.ene_all}
\title{Goodness-of-fit by number of exceedances}
\usage{
gof.ene_all(xdat, vecT = c(5, 10, 20, 40, 80), para = NULL, model = NULL)
}
\arguments{
\item{xdat}{Numeric vector of data.}

\item{vecT}{Vector of return periods.}

\item{para}{Parameter vector.}

\item{model}{Model name string.}
}
\value{
Sum of the relative absolute differences.
}
\description{
Internal function computing a goodness-of-fit measure
comparing the expected and observed numbers of exceedances of the
T-year levels for a set of return periods.
}
\keyword{internal}
