% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev.R
\name{init.gevmax}
\alias{init.gevmax}
\title{Initialize parameters for GEV MLE estimation}
\usage{
init.gevmax(data = NULL, ntry = NULL)
}
\arguments{
\item{data}{A numeric vector of data to be fitted.}

\item{ntry}{Number of initial parameter sets to generate.}
}
\value{
A matrix with `ntry` rows and 3 columns, where each row represents
a set of initial parameters (location, scale, shape) for the GEV distribution.
}
\description{
This function initializes parameters for GEV maximum likelihood estimation.
}
\details{
The function generates `ntry` sets of initial parameters for the GEV distribution.
It uses L-moment estimates as a starting point and then generates additional
sets of parameters using random perturbations.
}
\author{
Yonggwan Shin, Seokkap Ko, Jihong Park, Yire Shin, Jeong-Soo Park
}
\keyword{internal}
