% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{lmoms.md.park}
\alias{lmoms.md.park}
\title{Modified L-moments computation with failure handling}
\usage{
lmoms.md.park(x, nmom = 5, mtrim = FALSE, no.stop = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{nmom}{Number of L-moments.}

\item{mtrim}{If TRUE, uses left-trimmed TL-moments (leftrim=5).}

\item{no.stop}{If TRUE, returns \code{ifail=1} instead of stopping on error.}
}
\value{
A list of L-moments with an additional \code{ifail} element.
}
\description{
A slightly modified version of \code{lmomco::lmoms} (by
J. Park) which returns an \code{ifail} flag instead of stopping when
the L-moments cannot be computed.
}
\keyword{internal}
