% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev.R
\name{MS_pk}
\alias{MS_pk}
\title{Martins-Stedinger prior function}
\usage{
MS_pk(para = para, p = 6, q = 9)
}
\arguments{
\item{para}{A vector of GEV parameters (location, scale, shape).}

\item{p}{Shape parameter for beta distribution (default 6).}

\item{q}{Shape parameter for beta distribution (default 9).}
}
\value{
Prior probability value.
}
\description{
Martins-Stedinger prior function
}
\author{
Jeong-Soo Park
}
