% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev.R
\name{new_pf_norm}
\alias{new_pf_norm}
\title{Normal preference function for shape parameter}
\usage{
new_pf_norm(para = NULL, mu = NULL, std = NULL)
}
\arguments{
\item{para}{A vector of GEV parameters.}

\item{mu}{Mean for normal distribution.}

\item{std}{Standard deviation for normal distribution.}
}
\value{
Preference function value.
}
\description{
Normal preference function for shape parameter
}
\author{
Yonggwan Shin, Yire Shin, Jihong Park, Jeong-Soo Park
}
