% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{nllh.glme.gev11}
\alias{nllh.glme.gev11}
\title{GLME objective function for GEV11 model (mu0, sigma0, xi optimization)}
\usage{
nllh.glme.gev11(
  a,
  xdat = xdat,
  newtheta = newtheta,
  covinv = covinv,
  lcovdet = lcovdet,
  pen = pen,
  mu = mu,
  std = std,
  p = p,
  c0 = c0,
  c1 = c1,
  c2 = c2,
  q = q
)
}
\arguments{
\item{a}{Parameter vector (mu0, sigma0, xi).}

\item{xdat}{Data vector.}

\item{newtheta}{Full parameter vector (mu0, mu1, sigma0, sigma1, xi).}

\item{covinv}{Inverse covariance matrix of the sample L-moments.}

\item{lcovdet}{Log determinant of the covariance matrix.}

\item{pen}{Penalty type.}

\item{mu}{Normal penalty mean.}

\item{std}{Normal penalty std.}

\item{p}{Beta penalty shape.}

\item{c0}{Beta penalty support half-width.}

\item{c1}{Beta penalty scaling.}

\item{c2}{Beta penalty limit.}

\item{q}{Optional fixed second beta shape parameter.}
}
\value{
Penalized negative log-likelihood value.
}
\description{
GLME objective function for GEV11 model (mu0, sigma0, xi optimization)
}
\author{
Yonggwan Shin, Seokkap Ko, Jihong Park, Yire Shin, Jeong-Soo Park
}
\keyword{internal}
