% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PhliuAgromet}
\alias{PhliuAgromet}
\title{Phliu Agrometeorological Station Data}
\format{
A data frame with 40 rows and 9 columns:
\describe{
  \item{Station.ID}{Station identifier (character)}
  \item{year}{Year of observation (numeric, 1984-2023)}
  \item{prec}{Annual maximum daily precipitation in mm (numeric)}
  \item{Name}{Station name (character)}
  \item{zone}{Climate zone code (character)}
  \item{latitude}{Station latitude in degrees (numeric)}
  \item{longitude}{Station longitude in degrees (numeric)}
  \item{Starting.year}{Record start year (integer)}
  \item{Ending.year}{Record end year (numeric)}
}
}
\source{
Phliu Agrometeorological Station, Thailand.
}
\usage{
PhliuAgromet
}
\description{
Climate or meteorological data from the Phliu Agrometeorological Station
for extreme value analysis.
}
\examples{
data(PhliuAgromet)
head(PhliuAgromet)

}
\references{
Shin, Y., Shin, Y., Park, J., & Park, J. S. (2025). Generalized method of L-moment estimation
for stationary and nonstationary extreme value models. arXiv preprint arXiv:2512.20385.
\doi{10.48550/arXiv.2512.20385}
}
\keyword{datasets}
