% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PhliuAgromet}
\alias{PhliuAgromet}
\title{Phliu Agrometeorological Station Data}
\format{
A data frame containing meteorological measurements.
}
\source{
Phliu Agrometeorological Station.
}
\usage{
PhliuAgromet
}
\description{
Climate or meteorological data from the Phliu Agrometeorological Station
for extreme value analysis.
}
\examples{
data(PhliuAgromet)
head(PhliuAgromet)

}
\keyword{datasets}
