% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{pk.beta.ns}
\alias{pk.beta.ns}
\title{Beta preference function for non-stationary GEV}
\usage{
pk.beta.ns(para = NULL, lme.center = NULL, p = NULL, c0 = 0.3, c1 = 10, c2 = 5)
}
\arguments{
\item{para}{A vector of GEV parameters.}

\item{lme.center}{L-moment estimates as center.}

\item{p}{Shape parameter.}

\item{c0}{Limit parameter (default 0.3).}

\item{c1}{Scaling parameter (default 10).}

\item{c2}{Upper limit parameter (default 5).}
}
\value{
A list containing pk.one, p, and q.
}
\description{
Beta preference function for non-stationary GEV
}
\author{
Yonggwan Shin, Yire Shin, Jihong Park, Jeong-Soo Park
}
\keyword{internal}
