% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev11.R
\name{sel.para_all}
\alias{sel.para_all}
\title{Select best parameters based on GOF}
\usage{
sel.para_all(xdat, para.sel, model = NULL)
}
\description{
Select best parameters based on GOF
}
\keyword{internal}
