% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streamflow}
\alias{streamflow}
\title{Streamflow Data}
\format{
A data frame with 50 rows and 2 columns:
\describe{
  \item{Year}{Year of observation (character)}
  \item{r1}{Annual maximum streamflow value (numeric)}
}
}
\source{
Hydrological data for extreme value analysis.
}
\usage{
streamflow
}
\description{
Annual maximum streamflow measurements for extreme value analysis.
}
\examples{
data(streamflow)
head(streamflow)

}
\references{
Grego, J. M., Yates, P. A., & Mai, K. (2015). Standard error estimation for mixed flood distributions with historic maxima. Environmetrics, 26(3), 229-242. \doi{10.1002/env.2333}
Shin, Y., Shin, Y., & Park, J. S. (2025). Building nonstationary extreme value model using L-moments. Journal of the Korean Statistical Society, 1-24. \doi{10.1007/s42952-025-00325-3}
}
\keyword{datasets}
