% Generated by roxygen2: do not edit by hand
\name{streamflow}
\alias{streamflow}
\docType{data}
\title{Streamflow Data}
\description{
Streamflow measurement data for extreme value analysis.
}
\usage{data(streamflow)}
\format{
A data frame or vector containing streamflow measurements.
}
\keyword{datasets}
