% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strup.gev11.R
\name{strup.gev11}
\alias{strup.gev11}
\title{Weighted least squares estimation for the non-stationary GEV11 model}
\usage{
strup.gev11(xdat, init.rob = TRUE, wls.rob = FALSE, strup.fin.rob = TRUE)
}
\arguments{
\item{xdat}{A numeric vector of data to be fitted.}

\item{init.rob}{If TRUE (default), the initial location trend is fitted
by robust regression (\code{\link[robustbase]{lmrob}}); otherwise by
ordinary least squares.}

\item{wls.rob}{If TRUE, the log-scale regression uses robust regression
(default FALSE).}

\item{strup.fin.rob}{If TRUE (default), the final location trend
re-specification uses robust regression.}
}
\value{
A list containing:
\itemize{
  \item strup.para - Original WLS estimates (mu0, mu1, sigma0, sigma1, xi).
  \item strup.mdfy - Modified (re-specified) WLS estimates.
}
}
\description{
Estimates the parameters of the non-stationary GEV11 model by the
weighted least squares (WLS) method of Strupczewski and Kaczmarek (2001),
with the final-specification modification described in Shin et al.
(2025b, JKSS). The result is used as the pre-estimate (starting value)
by \code{\link{glme.gev11}} and \code{\link{lme.gev11}}.
}
\examples{
data(PhliuAgromet)
result <- strup.gev11(PhliuAgromet$prec)
print(result$strup.mdfy)

}
\references{
Strupczewski, W. G. & Kaczmarek, Z. (2001). Non-stationary approach to
at-site flood frequency modelling II. Weighted least squares estimation.
Journal of Hydrology, 248(1-4), 143-151.

Shin, Y., Shin, Y. & Park, J.-S. (2025). Building nonstationary extreme value
model using L-moments. Journal of the Korean Statistical Society, 54, 947-970.
\doi{10.1007/s42952-025-00325-3}
}
\seealso{
\code{\link{glme.gev11}}, \code{\link{lme.gev11}},
  \code{\link{GN16.gev11}}.
}
\author{
Yonggwan Shin, Seokkap Ko, Jihong Park, Yire Shin, Jeong-Soo Park
}
